+++ src/util/Document.js
@@ -24,6 +24,7 @@ module.exports = class Document {
     const json = { ...this };
     delete json._model;
     delete json._doc;
+    this._model.emit('save', this);
     return json;
   }
 };